refactor(gsx): extract pure text rules into GsxService.TextRules.cs
Pure code motion (private -> internal only) so tools/GsxTextProbe can
link the file directly, per the GsxProfileProbe convention.
