+++ MSFSBlindAssist/Services/GsxService.cs
@@ -26,7 +26,7 @@ namespace MSFSBlindAssist.Services;
 /// updates, and exposes events the UI form (AccessGSXForm) and MainForm
 /// background hook subscribe to.
 /// </summary>
-public sealed class GsxService : IDisposable
+public sealed partial class GsxService : IDisposable
 {
     // Distinct WM_USER message id — the main SimConnect uses 0x0402, this
     // one uses 0x0403 so both clients' ReceiveMessage calls are dispatched
@@ -53,14 +53,6 @@ public sealed class GsxService : IDisposable
     private static readonly TimeSpan TimerOnlyStatusAnnouncementInterval = TimeSpan.FromMinutes(30);
     private static readonly TimeSpan GroundConnectionTimerAnnouncementInterval = TimeSpan.FromMinutes(15);
     private static readonly TimeSpan FuelingProgressAnnouncementInterval = TimeSpan.FromSeconds(30);
-    private const int BoardingPassengerAnnouncementInterval = 10;
-    private const string CurrencyCodePattern =
-        @"(?:AED|AFN|ALL|AMD|ANG|AOA|ARS|AUD|AWG|AZN|BAM|BBD|BDT|BGN|BHD|BIF|BMD|BND|BOB|BOV|BRL|BSD|BTN|BWP|BYN|BZD|CAD|CDF|CHE|CHF|CHW|CLF|CLP|CNY|COP|COU|CRC|CUC|CUP|CVE|CZK|DJF|DKK|DOP|DZD|EGP|ERN|ETB|EUR|FJD|FKP|GBP|GEL|GHS|GIP|GMD|GNF|GTQ|GYD|HKD|HNL|HTG|HUF|IDR|ILS|INR|IQD|IRR|ISK|JMD|JOD|JPY|KES|KGS|KHR|KMF|KPW|KRW|KWD|KYD|KZT|LAK|LBP|LKR|LRD|LSL|LYD|MAD|MDL|MGA|MKD|MMK|MNT|MOP|MRU|MUR|MVR|MWK|MXN|MXV|MYR|MZN|NAD|NGN|NIO|NOK|NPR|NZD|OMR|PAB|PEN|PGK|PHP|PKR|PLN|PYG|QAR|RON|RSD|RUB|RWF|SAR|SBD|SCR|SDG|SEK|SGD|SHP|SLE|SLL|SOS|SRD|SSP|STN|SVC|SYP|SZL|THB|TJS|TMT|TND|TOP|TRY|TTD|TWD|TZS|UAH|UGX|USD|USN|UYI|UYU|UYW|UZS|VED|VES|VND|VUV|WST|XAF|XAG|XAU|XBA|XBB|XBC|XBD|XCD|XCG|XDR|XOF|XPD|XPF|XPT|XSU|XTS|XUA|XXX|YER|ZAR|ZMW|ZWG)";
-    private const string CurrencyWordPattern =
-        @"(?:ARIARY|BAHT|BALBOA|BIRR|BOLIVAR|BOLIVARES|BOLIVIANO|BOLIVIANOS|CEDI|CEDIS|COLON|COLONES|CORDOBA|CORDOBAS|DALASI|DENAR|DENARS|DINAR|DINARS|DIRHAM|DIRHAMS|DOLLAR|DOLLARS|DONG|DRAM|ESCUDO|ESCUDOS|EURO|EUROS|FLORIN|FORINT|FRANC|FRANCS|GOURDE|GOURDES|GUARANI|HRYVNIA|KINA|KIP|KORUNA|KORUNY|KRONA|KRONER|KRONOR|KWACHA|KWANZA|KYAT|LARI|LEI|LEK|LEMPIRA|LEMPIRAS|LEONE|LEV|LEVA|LILANGENI|LIRA|LIRE|LOTI|MANAT|METICAL|METICALS|NAIRA|NAKFA|OUGUIYA|PAANGA|PATACA|PATACAS|PESO|PESOS|POUND|POUNDS|PULA|QUETZAL|RAND|REAL|REALS|RIEL|RINGGIT|RIYAL|RIYALS|RMB|ROUBLE|ROUBLES|RUBLE|RUBLES|RUFIYAA|RUPEE|RUPEES|RUPIAH|SHEKEL|SHEKELS|SHILLING|SHILLINGS|SOL|SOM|SOMONI|TAKA|TALA|TENGE|TUGRIK|VATU|WON|YEN|YUAN|ZLOTY|ZLOTYS)";
-    private const string CurrencySymbolPattern = @"[$€£¥₩₹₽₺₪₫₴¢₦₱฿₡₲₵₭₮₾₼]";
-    private const string CurrencyTokenPattern =
-        @"(?:" + CurrencyCodePattern + @"|" + CurrencyWordPattern + @"|" + CurrencySymbolPattern + @")";
     private static readonly HashSet<string> AnnounceableReceiptFolders = new(StringComparer.OrdinalIgnoreCase)
     {
         "Catering",
@@ -1163,24 +1155,6 @@ public sealed class GsxService : IDisposable
         return string.Join(", ", changedParts);
     }
 
-    private static List<string> SplitTooltipParts(string text)
-    {
-        var parts = new List<string>();
-        if (string.IsNullOrWhiteSpace(text))
-            return parts;
-
-        foreach (string line in text.ReplaceLineEndings("\n").Split('\n'))
-        {
-            foreach (string segment in Regex.Split(line, @"(?<!\d),(?!\d)"))
-            {
-                string trimmed = segment.Trim();
-                if (trimmed.Length > 0)
-                    parts.Add(trimmed);
-            }
-        }
-        return parts;
-    }
-
     private static readonly Regex EtaSegmentRegex = new(
         @"\beta\b",
         RegexOptions.IgnoreCase | RegexOptions.Compiled);
@@ -1201,16 +1175,6 @@ public sealed class GsxService : IDisposable
             return text;
         return string.Join(", ", kept);
     }
-
-    // Pax-count-only segments (the things we silence when boarding's
-    // milestone hasn't advanced). Whole-segment match so any segment that
-    // carries additional info ("rear loader leaving while 5 boarded") is
-    // left alone — only standalone "pax 5/100" / "5 passengers" / "5 pax
-    // boarded" type lines, plus GSX's "Passenger boarding 5/100 passengers"
-    // status shape, are stripped.
-    private static readonly Regex PaxOnlySegmentRegex = new(
-        @"^\s*(?:\[gsx\]\s+)?(?:(?:passenger\s+(?:de)?boarding)\s+)?(?:pax\s+\d{1,4}(?:\s*/\s*\d{1,4})?|\d{1,4}(?:\s*/\s*\d{1,4})?\s+(?:passengers?|pax)(?:\s+boarded)?)\s*$",
-        RegexOptions.IgnoreCase | RegexOptions.Compiled);
 
     // Aircraft fuel quantity is GSX's rapidly-updating internal aircraft
     // fuel row, separate from the useful fueling progress announcement.
@@ -1291,9 +1255,6 @@ public sealed class GsxService : IDisposable
 
         return StripSegmentsMatching(announceText, PaxOnlySegmentRegex);
     }
-
-    private static bool IsBoardingAnnouncementBoundary(int passengers) =>
-        passengers <= 1 || passengers % BoardingPassengerAnnouncementInterval == 0;
 
     private string StripThrottledBagsSegments(string announceText, string fullText)
     {
@@ -1423,21 +1384,6 @@ public sealed class GsxService : IDisposable
             @"\bbill[\s:]*\$?\d+(?:[.,]\d+)?\b",
             RegexOptions.IgnoreCase);
 
-    private static string StripSegmentsMatching(string text, Regex regex)
-    {
-        var parts = SplitTooltipParts(text);
-        if (parts.Count == 0)
-            return text;
-
-        var kept = parts.Where(p => !regex.IsMatch(p)).ToList();
-        if (kept.Count == 0)
-            return string.Empty;
-        if (kept.Count == parts.Count)
-            return text;
-        return string.Join(", ", kept);
-    }
-
-
     private static bool IsPlaceholderLiveServiceText(string text)
     {
         string normalized = NormalizeWhitespace(text);
@@ -1556,20 +1502,6 @@ public sealed class GsxService : IDisposable
         return false;
     }
 
-    // Boarding-progress milestones — chosen so the user hears:
-    //   * pax 0 (service started, nobody on yet)
-    //   * pax 1 (boarding has actually begun)
-    //   * every multiple of BoardingPassengerAnnouncementInterval (10, 20, …)
-    // with no upper cap, so 110 / 120 / 130 / … keep announcing instead of
-    // collapsing into a single milestone-100 ceiling like Tower's original
-    // formula did.
-    private static int ComputeBoardingMilestone(int passengers)
-    {
-        if (passengers <= 0) return 0;
-        if (passengers < BoardingPassengerAnnouncementInterval) return 1;
-        return (passengers / BoardingPassengerAnnouncementInterval) + 1;
-    }
-
     // Variant used by the progress throttles (fueling, boarding). Always uses
     // the fixed service prefix as the key root, optionally suffixed with the
     // operator if one is extractable. This guarantees the dict key stays the
@@ -1584,107 +1516,10 @@ public sealed class GsxService : IDisposable
             : $"{fixedService}|{serviceOperator}";
     }
 
-    private static bool TryParsePassengerCount(string text, out int passengers)
-    {
-        passengers = 0;
-
-        // GSX writes the boarding count several ways. The "N/M" forms must
-        // be tried first — the bare "\b\d{1,3}\s*passengers\b" pattern
-        // would otherwise greedily capture M (the total) out of "50/550
-        // passengers", poisoning the milestone throttle by recording
-        // milestone 100 (M/10) and silencing every subsequent milestone
-        // below that for the rest of the session.
-        var match = Regex.Match(
-            text,
-            @"\bpassenger\s+(?:de)?boarding\s+(?<count>\d{1,4})\s*/\s*\d+\s*(?:passengers|pax)\b",
-            RegexOptions.IgnoreCase);
-        if (!match.Success)
-        {
-            match = Regex.Match(
-                text,
-                @"\bpassenger\s+(?:de)?boarding\s+(?<count>\d{1,4})\s*(?:passengers|pax)\b",
-                RegexOptions.IgnoreCase);
-        }
-        if (!match.Success)
-        {
-            match = Regex.Match(
-                text,
-                @"\b(?<count>\d{1,4})\s*/\s*\d+\s*(?:passengers|pax)\b",
-                RegexOptions.IgnoreCase);
-        }
-        if (!match.Success)
-        {
-            match = Regex.Match(
-                text,
-                @"\bpax\s+(?<count>\d{1,4})\s*/\s*\d+\b",
-                RegexOptions.IgnoreCase);
-        }
-        if (!match.Success)
-        {
-            match = Regex.Match(
-                text,
-                @"\b(?<count>\d{1,4})\s*(?:passengers|pax)\b",
-                RegexOptions.IgnoreCase);
-        }
-        if (!match.Success)
-        {
-            match = Regex.Match(
-                text,
-                @"\bpax\s+(?<count>\d{1,4})\b",
-                RegexOptions.IgnoreCase);
-        }
-        if (!match.Success)
-            return false;
-
-        passengers = Math.Clamp(int.Parse(match.Groups["count"].Value, CultureInfo.InvariantCulture), 0, 999);
-        return true;
-    }
-
     private static TimeSpan GetTimerOnlyAnnouncementInterval(string text) =>
         IsGroundConnectionService(text)
             ? GroundConnectionTimerAnnouncementInterval
             : TimerOnlyStatusAnnouncementInterval;
-
-    private static bool IsTimerStatusText(string text) =>
-        text.Contains("timer:", StringComparison.OrdinalIgnoreCase);
-
-    private static bool TimerStatusContextEquals(string currentText, string previousText)
-    {
-        if (string.IsNullOrWhiteSpace(previousText))
-            return false;
-
-        string currentContext = BuildTimerStatusContextKey(currentText);
-        string previousContext = BuildTimerStatusContextKey(previousText);
-        return currentContext.Length > 0
-            && string.Equals(currentContext, previousContext, StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static string BuildTimerStatusContextKey(string text)
-    {
-        var parts = SplitTooltipParts(text);
-        if (parts.Count == 0)
-            return string.Empty;
-
-        var stableParts = parts
-            .Select(BuildTimerStatusContextPart)
-            .Where(part => !string.IsNullOrWhiteSpace(part));
-
-        return string.Join("|", stableParts);
-    }
-
-    private static string BuildTimerStatusContextPart(string part)
-    {
-        if (part.Equals("Current charges:", StringComparison.OrdinalIgnoreCase))
-            return string.Empty;
-
-        if (IsTimerStatusLine(part))
-            return NormalizeStatusStableText(FormatGroundConnectionTimerServiceText(part));
-
-        if (IsChargeStatusLine(part))
-            return string.Empty;
-
-        return NormalizeStatusStableText(part);
-    }
 
     private void ReloadAndPublishSettings()
     {
@@ -2411,9 +2246,6 @@ public sealed class GsxService : IDisposable
     private static string DecodeHtml(string value) =>
         System.Net.WebUtility.HtmlDecode(value ?? string.Empty);
 
-    private static string NormalizeWhitespace(string value) =>
-        Regex.Replace(value.ReplaceLineEndings(" "), @"\s+", " ").Trim();
-
     // ─────────────────────────────────────────────────────────────────────
     // Status text + event raise helpers.
     // ─────────────────────────────────────────────────────────────────────
@@ -2747,42 +2579,10 @@ public sealed class GsxService : IDisposable
     private static bool HasConnectedStatusWord(string text) =>
         ConnectedWordRegex.IsMatch(text) && !DisconnectedWordRegex.IsMatch(text);
 
-    private static bool IsChargeStatusLine(string text)
-    {
-        string normalized = text.ToLowerInvariant();
-        return normalized.Contains("timer:")
-            || normalized.Contains("invoice:")
-            || normalized.Contains("eur ")
-            || normalized.Contains("â‚¬")
-            || normalized.Contains("$")
-            || normalized.Contains("£")
-            || ContainsMoneyAmount(text);
-    }
-
-    private static bool ContainsMoneyAmount(string text) =>
-        Regex.IsMatch(
-            text,
-            $@"{CurrencyTokenPattern}\s*\d|\d[\d,.]*\s*{CurrencyTokenPattern}",
-            RegexOptions.IgnoreCase);
-
-    private static bool IsTimerStatusLine(string text) =>
-        text.Contains("timer:", StringComparison.OrdinalIgnoreCase);
-
     private static bool IsRunningGroundConnectionTimerLine(string text) =>
         IsTimerStatusLine(text)
         && text.Contains("running", StringComparison.OrdinalIgnoreCase)
         && IsGroundConnectionService(text);
-
-    private static string FormatGroundConnectionTimerServiceText(string timerLine)
-    {
-        string serviceName = Regex.Replace(timerLine, @"\s+timer\s*:.*$", string.Empty,
-            RegexOptions.IgnoreCase);
-        serviceName = NormalizeWhitespace(serviceName);
-        if (string.IsNullOrWhiteSpace(serviceName))
-            return string.Empty;
-
-        return $"{serviceName} service is running";
-    }
 
     private static string FormatCompletedServiceTotal(string serviceText, IReadOnlyList<string> chargeRows)
     {
@@ -3277,47 +3077,6 @@ public sealed class GsxService : IDisposable
             yield return row;
             previous = row;
         }
-    }
-
-    private static string NormalizeStatusStableText(string text)
-    {
-        if (string.IsNullOrWhiteSpace(text))
-            return string.Empty;
-
-        string stable = text;
-        stable = Regex.Replace(stable, @"\b\d{1,2}:\d{2}(?::\d{2})?\b", "<time>");
-        // Bucket durations into 5-minute groups so an ETA that counts down
-        // second-by-second doesn't re-announce on every tick. A re-announce
-        // fires only when the value crosses a 5-minute boundary (which
-        // approximates the user-requested "tell me again if the ETA changes
-        // by ~5 minutes" behaviour).
-        stable = DurationTokenRegex.Replace(stable, BucketDurationToken);
-        stable = Regex.Replace(stable, $@"{CurrencyTokenPattern}\s*\d[\d,.]*|\d[\d,.]*\s*{CurrencyTokenPattern}", "<price>",
-            RegexOptions.IgnoreCase);
-        stable = Regex.Replace(stable, @"\(~?\s*<price>\)", "(<price>)",
-            RegexOptions.IgnoreCase);
-        return NormalizeWhitespace(stable);
-    }
-
-    private static readonly Regex DurationTokenRegex = new(
-        @"\b(?<num>\d+(?:[.,]\d+)?)\s*(?<unit>seconds?|secs?|minutes?|mins?)\b",
-        RegexOptions.IgnoreCase | RegexOptions.Compiled);
-
-    private static string BucketDurationToken(Match match)
-    {
-        if (!double.TryParse(
-                match.Groups["num"].Value.Replace(',', '.'),
-                System.Globalization.NumberStyles.Float,
-                System.Globalization.CultureInfo.InvariantCulture,
-                out double n))
-        {
-            return "<duration>";
-        }
-
-        string unit = match.Groups["unit"].Value.ToLowerInvariant();
-        double totalSeconds = unit.StartsWith("m") ? n * 60.0 : n;
-        int bucket = (int)Math.Floor(totalSeconds / 300.0);
-        return $"<duration-{bucket}>";
     }
 
     private void UpdateStatusText()
